requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
colorama==0.4.6
coredis==4.16.0
Deprecated==1.2.14
Flask==3.0.0
helper==2.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
paho-mqtt==1.6.1
psycopg2==2.9.9
pymodbus==3.6.2
Pympler==1.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
wrapt==1.16.0
